healthcare_system/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
google-generativeai==0.8.6
requests==2.33.1
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
psycopg[binary]==3.2.9
